fix: add validations to RBAC services

+++ apps/auth-service/src/services/permission.service.ts
@@ -8,15 +8,30 @@ const getAllPermission = async (): Promise<string[]> => {
 };
 
 const getPermissionById = async (permissionId: string) => {
+  if (!permissionId) throw new Error('Missing permission ID');
+
   const permission = await Permission.findOne({ where: { id: permissionId } });
-  return permission?.permission;
+  if (!permission) throw new Error('Permission not found');
+
+  return permission.permission;
 };
 
 const createPermission = async (permission: string) => {
+  if (!permission) throw new Error('Missing required field');
+
+  // Check if permission already exists
+  const existingPermission = await Permission.findOne({ where: { permission } });
+  if (existingPermission) throw new Error('Permission already exists');
+
   await Permission.create({ permission });
 };
 
 const deletePermission = async (permissionId: string) => {
+  if (!permissionId) throw new Error('Missing permission ID');
+
+  const permission = await Permission.findOne({ where: { id: permissionId } });
+  if (!permission) throw new Error('Permission not found');
+
   await Permission.destroy({ where: { id: permissionId } });
 };
